backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
aif360==0.6.1
python-dotenv==1.0.1
httpx==0.27.2
pdfkit==1.0.0
reportlab==4.2.2
supabase==2.6.0
asyncpg==0.29.0
sqlalchemy==2.0.31
imbalanced-learn==0.12.3
joblib==1.4.2
shap==0.45.1
lime==0.2.0.1
scipy==1.13.1
statsmodels==0.14.2
slowapi==0.1.9
google-generativeai>=0.4.0
